+++ sources/entities/object/object.hpp
@@ -11,20 +11,20 @@ protected:
     double _x = 0;
     double _y = 0;
 
-    int _sizeX;
-    int _sizeY;
+    double _sizeX;
+    double _sizeY;
 
 public:
     MapEncoding type;
 
     Object(){};
-    Object(int ix, int iy, int sizeX, int sizeY, MapEncoding type);
+    Object(int ix, int iy, double sizeX, double sizeY, MapEncoding type);
     virtual ~Object() = default;
 
     auto GetShape() -> Shape;
 };
 
-Object::Object(int ix, int iy, int sizeX, int sizeY, MapEncoding type) : _x(multiply(ix, DELTA_X)), _y(multiply(HEIGHT - iy - 1, DELTA_Y)), _sizeX(sizeX), _sizeY(sizeY), type(type) {}
+Object::Object(int ix, int iy, double sizeX, double sizeY, MapEncoding type) : _x(multiply(ix, DELTA_X)), _y(multiply(HEIGHT - iy - 1, DELTA_Y)), _sizeX(sizeX), _sizeY(sizeY), type(type) {}
 
 inline auto Object::GetShape() -> Shape
 {
